src/main/java/nl/hu/cisq1/lingo/words/domain/Word.java
package nl.hu.cisq1.lingo.words.domain;

import org.apache.commons.lang3.StringUtils;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;

@Entity(name = "words")
public class Word {
    @Id
    @Column(name = "word")
    private String value;
    private Integer length;

    public Word() {}
    public Word(String word) {
        if(StringUtils.isBlank(word)) {
            throw new IllegalArgumentException("woord mag niet leeg zijn");
        }

        this.value = word;
        this.length = word.length();
    }

    public String getValue() {
        return value;
    }

    public Integer getLength() {
        return length;
    }
}
